+++ player/main.c
@@ -472,9 +472,7 @@ int mp_initialize(struct MPContext *mpctx, char **options)
     if (opts->force_vo == 2 && handle_force_window(mpctx, false) < 0)
         return -1;
 
-#if !defined(__MINGW32__)
     mpctx->ipc_ctx = mp_init_ipc(mpctx->clients, mpctx->global);
-#endif
 
 #ifdef _WIN32
     if (opts->w32_priority > 0)
